requirements.txt
telethon>=1.34.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
httpx>=0.25.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pydub>=0.25.1
openai-whisper>=20231117
fastapi>=0.110.0
uvicorn>=0.27.0
